# packages/sdk/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/me/watchlists": {
      "get": {
        "operationId": "WatchlistsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "The user’s watchlists (created with the Majors and Global defaults on first read).",
        "tags": [
          "terminal"
        ]
      },
      "post": {
        "operationId": "WatchlistsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "pattern": "^[\\p{L}\\p{N} ._-]{1,40}$"
                  },
                  "symbols": {
                    "default": [],
                    "maxItems": 500,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[A-Z0-9][A-Z0-9._-]{0,31}$"
                    }
                  }
                },
                "required": [
                  "name"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a watchlist (≤ 20 per user, ≤ 500 symbols each).",
        "tags": [
          "terminal"
        ]
      }
    },
    "/me/watchlists/{id}": {
      "put": {
        "operationId": "WatchlistsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "pattern": "^[\\p{L}\\p{N} ._-]{1,40}$"
                  },
                  "symbols": {
                    "maxItems": 500,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[A-Z0-9][A-Z0-9._-]{0,31}$"
                    }
                  },
                  "position": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1000
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Rename, reorder symbols (drag or add/remove) or move a watchlist.",
        "tags": [
          "terminal"
        ]
      },
      "delete": {
        "operationId": "WatchlistsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete a watchlist.",
        "tags": [
          "terminal"
        ]
      }
    },
...
  },
  "info": {
    "title": "KORA API",
    "description": "KORA platform API (PAPER only). Money and prices are decimal strings.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kora_at"
      }
    },
...
  }
}